Semester 2/COMP7310 - Artificial intelligence of things/project/backend/requirements.txt
Flask
Flask-SocketIO
paho-mqtt
sqlite3